02-equipment-hours-validation/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.5
